feat(storybook): migrate to Storybook 9 + Vite 7 with pure ESM config\n\n- Bump storybook and vite to 9.x/7.x and add @storybook/react-vite + builder-vite\n- Remove deprecated addon-essentials; keep links, docs, a11y\n- Convert Storybook config to ESM-only: add main.mjs, lazy-load vite-tsconfig-paths\n- Add JS version of production-stories config; stop importing TS at runtime\n- Clean overrides; remove esbuild-register pins\n- Verify production build works (apps/storybook/storybook-static)\n\nCo-authored-by: Tulio Pinheiro Cunha <contact@tuliocunha.dev>
