+++ apps/storybook/package.json
@@ -3,11 +3,10 @@
   "private": true,
   "type": "module",
   "devDependencies": {
-    "storybook": "^8.6.14",
-    "@storybook/react-vite": "^8.6.14",
-    "@storybook/builder-vite": "^8.6.14",
-    "vite": "^5.4.10",
-    "semver": "^7.6.0"
+    "storybook": "^9.1.1",
+    "@storybook/react-vite": "^9.1.1",
+    "@storybook/builder-vite": "^9.1.1",
+    "vite": "^7.1.1"
   },
   "scripts": {
     "dev": "storybook dev -p 6006 -c .storybook",
