+++ package.json
@@ -31,10 +31,9 @@
     "@axe-core/react": "^4.10.2",
     "@jest/globals": "^30.0.5",
     "@playwright/test": "^1.54.2",
-    "@storybook/addon-essentials": "^8.6.14",
-    "@storybook/addon-a11y": "^8.6.14",
-    "@storybook/addon-docs": "^8.6.14",
-    "@storybook/addon-links": "^8.6.14",
+    "@storybook/addon-a11y": "^9.1.1",
+    "@storybook/addon-docs": "^9.1.1",
+    "@storybook/addon-links": "^9.1.1",
     "@testing-library/jest-dom": "^6.6.4",
     "@testing-library/react": "^16.3.0",
     "@testing-library/user-event": "^14.6.1",
@@ -53,7 +52,9 @@
     "react-dom": "^19.1.1",
     "rollup-plugin-visualizer": "^6.0.3",
     "serve": "^14.2.4",
-    "storybook": "^8.4.6",
+    "@storybook/react-vite": "^9.1.1",
+    "@storybook/builder-vite": "^9.1.1",
+    "storybook": "^9.1.1",
     "tailwindcss": "^4.1.11",
     "tailwindcss-filters": "^3.0.0",
     "typescript": "^5.9.2",
@@ -61,7 +62,7 @@
     "vite-plugin-lightningcss": "^0.0.5",
     "vite-tsconfig-paths": "^5.1.4",
     "vitest": "^3.2.4",
-    "vite": "^5.4.10"
+    "vite": "^7.1.1"
   },
   "engines": {
     "bun": ">=1.0.0"
@@ -121,10 +122,7 @@
   "overrides": {
     "@stoplight/spectral-core": "1.19.3",
     "@stoplight/spectral-functions": "1.9.2",
-    "esbuild-register": "3.5.0",
-    "vite": "5.4.10",
-    "storybook": "8.6.14",
-    "@storybook/*": "8.6.14"
+    /* Removed duplicate and redundant `esbuild-register` overrides */
   },
   "scripts": {
     "build": "bun run build:lib && bun run build:storybook",
